wip: peps tests BoundaryMps measure
