+++ yastn/tn/fpeps/__init__.py
@@ -4,10 +4,9 @@ from ._initialize import product_peps, load_from_dict
 from ._evolution import evolution_step_
 from ._doublePepsTensor import DoublePepsTensor
 from .gates import Gates, Gate_local, Gate_nn
-# from ._gates_auxlliary import apply_gate_onsite
+from ._gates_auxlliary import apply_gate_onsite
 from ._ctmrg import ctmrg_
 from .envs._env_ctm import EnvCTM
 from .envs._env_ntu import EnvNTU
 from .envs._env_boundary_mps import EnvBoundaryMps
 from .envs._env_cluster_approximate import EnvApproximate
-from .envs._measure import measure_1site, measure_2site, sample, sample_MC_
